+++ web/packages/pyxis-app/src/components/organisms/index.ts
@@ -8,7 +8,7 @@ export * from './DiscordMappingPanel';
 export * from './FlyerField';
 export * from './NewShowModal';
 export * from './Panel';
-export * from './PostShowLog';
+export * from './ShowLog';
 export * from './Settings';
 export * from './ShowDetail';
 export * from './Shows';